requirements.txt
faiss-cpu
sentence-transformers
pymupdf
transformers
pillow
groq
gradio
numpy
python-dotenv